ma-poca/mapoca/pyproject.toml
[project]
name = "mapoca"
version = "0.1.0"
description = "Unity Machine Learning Agents, MA-POCA"
authors = [{ "name" = "Unity Technologies", email = "ML-Agents@unity3d.com" }]
readme = "README.md"
requires-python = ">=3.10.1,<3.10.12"
dependencies = [
    "grpcio>=1.11.0",
    "h5py>=2.9.0",
    "mlagents_envs==1.1.0",
    "mlagents==1.1.0",
    "numpy>=1.23.5,<1.24.0",
    "Pillow>=4.2.1",
    "protobuf>=3.6",
    "pyyaml>=3.1.0",
    "pyglet",
    "torch>=2.0",
    # Windows ver. of PyTorch doesn't work from PyPi. Installation:
    # https://github.com/Unity-Technologies/ml-agents/blob/release_18_docs/docs/Installation.md#windows-installing-pytorch
    # Torch only working on python 3.9 for 1.8.0 and above. Details see:
    # https://github.com/pytorch/pytorch/issues/50014
    # "torch>=1.8.0,<1.9.0;(platform_system!='Windows' and python_version>='3.9')",
    # "torch>=1.6.0,<1.9.0;(platform_system!='Windows' and python_version<'3.9')",
    "tensorboard>=1.15",
    # cattrs 1.1.0 dropped support for python 3.6, but 1.0.0 doesn't work for python 3.9
    # Since there's no version that supports both, we have to draw the line somwehere.
    # "cattrs<1.1.0; python_version<'3.8'",
    "cattrs<=1.5",
    # "cattrs>=1.1.0,<1.7; python_version>='3.8'",
    "attrs>=19.3.0",
    'pypiwin32==223;platform_system=="Windows"',
    # "importlib_metadata; python_version<'3.8'",
    # Dependencies for particles envs
    "pettingzoo[mpe]>=1.0",
    # "gym==0.10.5",
    # "multiagent @ git+https://github.com/openai/multiagent-particle-envs.git",
]

[project.urls]
homepage = "https://github.com/Unity-Technologies/paper-ml-agents"
repository = "https://github.com/Unity-Technologies/paper-ml-agents"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mapoca-learn = "mapoca.trainers.learn:main"
mapoca-run-experiment = "mapoca.trainers.run_experiment:main"

[project.entry-points.'mapoca.stats_writer']
default = "mapoca.plugins.stats_writer:get_default_stats_writers"
